+++ Runtime/Extensions/MonoBehaviourExtensions.cs
@@ -6,7 +6,6 @@ namespace Packages.UniKit.Runtime.Extensions
     public static class MonoBehaviourExtensions
     {
         public static TComponent GetOrThrow<TComponent>(this MonoBehaviour monoBehaviour)
-            where TComponent : Component
         {
             if (ReferenceEquals(monoBehaviour, null))
             {