+++ Tests/Runtime/Tests_Extensions/Tests_GameObjectExtensions.cs
@@ -28,6 +28,23 @@ namespace Packages.UniKit.Tests.Runtime.Tests_Extensions
         }
 
         [UnityTest]
+        public IEnumerator GetOrThrow_WITH_ValidInterface_SHOULD_ReturnInterface()
+        {
+            GameObject emptyPrefab = new GameObject("TestGameObject");
+            var gameObj = Object.Instantiate(emptyPrefab);
+            gameObj.AddComponent<BazComponent>();
+
+            yield return null;
+
+            var baz = gameObj.GetOrThrow<IBaz>();
+
+            yield return null;
+
+            Assert.IsNotNull(baz);
+            Assert.AreEqual(typeof(BazComponent), baz.GetType());
+        }
+
+        [UnityTest]
         public IEnumerator GetOrThrow_WITH_InvalidComponent_SHOULD_Throw()
         {
             GameObject emptyPrefab = new GameObject("TestGameObject");
@@ -36,6 +53,17 @@ namespace Packages.UniKit.Tests.Runtime.Tests_Extensions
             yield return null;
 
             Assert.Throws<ArgumentException>(() => gameObj.GetOrThrow<FooComponent>());
+        }
+
+        [UnityTest]
+        public IEnumerator GetOrThrow_WITH_InvalidInterface_SHOULD_Throw()
+        {
+            GameObject emptyPrefab = new GameObject("TestGameObject");
+            var gameObj = Object.Instantiate(emptyPrefab);
+            
+            yield return null;
+
+            Assert.Throws<ArgumentException>(() => gameObj.GetOrThrow<IBaz>());
         }
 
         [UnityTest]